fix: rename session validator
